pyproject.toml
[tool.poetry]
name = "stackr"
version = "0.1.0"
description = ""
authors = ["Dennis Gera <geradennis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
fastapi = "^0.115.4"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
pandas = "^2.2.3"
plotly = "^5.24.1"
alembic = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
